% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_and_merge.R
\name{add_and_merge}
\alias{add_and_merge}
\title{Add content to spreadsheets}
\usage{
add_and_merge(value, sheet, row_num, colIndex, max_col, style = column_title)
}
\arguments{
\item{value}{the value to add}

\item{sheet}{sheet name}

\item{row_num}{cell row number}

\item{colIndex}{minimum column}

\item{max_col}{maximum column}

\item{style}{formatting style}
}
\description{
Similar to add_stuff but will also merge columns
}
