% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_back.R
\name{move_back}
\alias{move_back}
\title{Adjust quarters}
\usage{
move_back(value, move)
}
\arguments{
\item{value}{input value between 1 and 4}

\item{move}{the number of periods back from current}
}
\description{
Used to adjust quarters/periods for MOU lists. Very simple
}
