% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{rnd_perc}
\alias{rnd_perc}
\title{Round data}
\usage{
rnd_perc(x, digits = 0)
}
\arguments{
\item{x}{data}
}
\description{
Round decimal point number i.e. 0.9 to percentage i.e. 90\%, always rounds 0.5 up
}
